+++ visualization/app/codeCharta/state/store/dynamicSettings/colorRange/colorRange.reducer.spec.ts
@@ -3,13 +3,13 @@ import { defaultColorRange, setColorRange } from "./colorRange.actions"
 
 describe("colorRange", () => {
 	it("should set new colorRange", () => {
-		const result = colorRange({ from: 0, to: 0, min: 0, max: 0 }, setColorRange({ from: 21, to: 42, min: 0, max: 100 }))
+		const result = colorRange({ from: 0, to: 0 }, setColorRange({ from: 21, to: 42 }))
 
 		expect(result).toEqual({ from: 21, to: 42, min: 0, max: 100 })
 	})
 
 	it("should set default colorRange", () => {
-		const result = colorRange({ from: 33, to: 66, min: 0, max: 10 }, setColorRange())
+		const result = colorRange({ from: 33, to: 66 }, setColorRange())
 
 		expect(result).toEqual(defaultColorRange)
 	})